Split form section in a new component
